+++ src/azure-cli/azure/cli/command_modules/acr/_params.py
@@ -583,6 +583,8 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
                    help='Indicate whether garbage collection is enabled. It is enabled by default.', arg_type=get_three_state_flag(), required=False, default="true")
         c.argument('garbage_collection_schedule', options_list=['--gc-schedule'],
                    help='Used to determine garbage collection schedule. Uses cron expression to determine the schedule. If not specified, garbage collection is set to run once a day.', required=False, default="0 0 * * *")
+        c.argument('identity', options_list=['--identity'], help='Resource ID of a user-assigned managed identity to authenticate the connected registry with its parent. Required when --auth-type is ManagedIdentity.')
+        c.argument('auth_type', arg_type=get_enum_type(['SyncToken', 'ManagedIdentity']), options_list=['--auth-type'], help='Authentication type used by the connected registry to sync with its parent. Defaults to SyncToken.')
 
     with self.argument_context('acr connected-registry update') as c:
         c.argument('log_level', help='Set the log level for logging on the instance. Accepted log levels are Debug, Information, Warning, Error, and None.')
@@ -600,6 +602,9 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('garbage_collection_enabled', options_list=['--gc-enabled'],
                    help='Indicate whether garbage collection is enabled. It is enabled by default.', arg_type=get_three_state_flag())
         c.argument('garbage_collection_schedule', options_list=['--gc-schedule'], help='Used to determine garbage collection schedule. Uses cron expression to determine the schedule. If not specified, garbage collection is set to run once a day.')
+        c.argument('identity', options_list=['--identity'], help='Resource ID of a user-assigned managed identity. Required when migrating --auth-type to ManagedIdentity.')
+        c.argument('auth_type', arg_type=get_enum_type(['SyncToken', 'ManagedIdentity']), options_list=['--auth-type'], help='Target authentication type. Use to migrate between SyncToken and ManagedIdentity. Same-mode credential rotation is not supported and connected registry must be in Offline state.')
+        c.argument('sync_token_name', options_list=['--sync-token'], help='Existing sync token name. Required when migrating --auth-type back to SyncToken.')
     with self.argument_context('acr connected-registry permissions') as c:
         c.argument('add_repos', options_list=['--add'], nargs='*',
                    help='repository permissions to be added to the targeted connected registry and it\'s ancestors sync scope maps. Use the format "--add [REPO1 REPO2 ...]" per flag. ' + repo_valid_actions)